UpdateAccounting before locking more tokens

+++ test/helper.js
@@ -6,7 +6,7 @@ const AMPL_DECIMALS = 9;
 
 function $AMPL (x) {
   const ordinate = new BN(10 ** AMPL_DECIMALS);
-  return new BN(parseInt(x)).mul(ordinate);
+  return new BN(parseFloat(x)).mul(ordinate);
 }
 
 async function invokeRebase (ampl, perc) {
